+++ conn/mark_default.go
@@ -5,7 +5,7 @@
  * Copyright (C) 2017-2019 WireGuard LLC. All Rights Reserved.
  */
 
-package device
+package conn
 
 func (bind *nativeBind) SetMark(mark uint32) error {
 	return nil
